Move emitters to their own file
